feat: Protegiendo rutas y dando accesos solo al administrador a estados y estados tecnicos a través de middleware

+++ routes/web.php
@@ -24,20 +24,20 @@ Route::get('/pantallas', [PantallaController::class, 'index'])->middleware(['aut
 Route::get('/pantallas/{pantalla:orden_servicio}/show', [PantallaController::class, 'show'])->middleware(['auth', 'verified'])->name('pantallas.show');
 Route::get('/pantallas/{pantalla:orden_servicio}/edit', [PantallaController::class, 'edit'])->middleware(['auth', 'verified'])->name('pantallas.edit');
 
-Route::get('/estados', [EstadoController::class, 'index'])->middleware(['auth', 'verified'])->name('estados.index');
-Route::get('estados/create', [EstadoController::class, 'create'])->middleware(['auth', 'verified'])->name('estados.create');
-Route::post('/estados', [EstadoController::class, 'store'])->middleware(['auth', 'verified'])->name('estados.store');
-Route::get('/estados/{estado}/edit', [EstadoController::class, 'edit'])->middleware(['auth', 'verified'])->name('estados.edit');
-Route::put('/estados/{estado}', [EstadoController::class, 'update'])->middleware(['auth', 'verified'])->name('estados.update');
-Route::delete('/estados/{estado}', [EstadoController::class, 'destroy'])->middleware(['auth', 'verified'])->name('estados.destroy');
+Route::get('/estados', [EstadoController::class, 'index'])->middleware(['auth', 'verified', 'rol.admin'])->name('estados.index');
+Route::get('estados/create', [EstadoController::class, 'create'])->middleware(['auth', 'verified', 'rol.admin'])->name('estados.create');
+Route::post('/estados', [EstadoController::class, 'store'])->middleware(['auth', 'verified', 'rol.admin'])->name('estados.store');
+Route::get('/estados/{estado}/edit', [EstadoController::class, 'edit'])->middleware(['auth', 'verified', 'rol.admin'])->name('estados.edit');
+Route::put('/estados/{estado}', [EstadoController::class, 'update'])->middleware(['auth', 'verified', 'rol.admin'])->name('estados.update');
+Route::delete('/estados/{estado}', [EstadoController::class, 'destroy'])->middleware(['auth', 'verified', 'rol.admin'])->name('estados.destroy');
 
 
-Route::get('/estados_tecnicos', [EstadoTecnicoController::class, 'index'])->middleware(['auth', 'verified'])->name('estados_tecnicos.index');
-Route::get('estados_tecnicos/create', [EstadoTecnicoController::class, 'create'])->middleware(['auth', 'verified'])->name('estados_tecnicos.create');
-Route::post('/estados_tecnicos', [EstadoTecnicoController::class, 'store'])->middleware(['auth', 'verified'])->name('estados_tecnicos.store');
-Route::get('/estados_tecnicos/{estado_tecnico}/edit', [EstadoTecnicoController::class, 'edit'])->middleware(['auth', 'verified'])->name('estados_tecnicos.edit');
-Route::put('/estados_tecnicos/{estado_tecnico}', [EstadoTecnicoController::class, 'update'])->middleware(['auth', 'verified'])->name('estados_tecnicos.update');
-Route::delete('/estados_tecnicos/{estado_tecnico}', [EstadoTecnicoController::class, 'destroy'])->middleware(['auth', 'verified'])->name('estados_tecnicos.destroy');
+Route::get('/estados_tecnicos', [EstadoTecnicoController::class, 'index'])->middleware(['auth', 'verified', 'rol.admin'])->name('estados_tecnicos.index');
+Route::get('estados_tecnicos/create', [EstadoTecnicoController::class, 'create'])->middleware(['auth', 'verified', 'rol.admin'])->name('estados_tecnicos.create');
+Route::post('/estados_tecnicos', [EstadoTecnicoController::class, 'store'])->middleware(['auth', 'verified', 'rol.admin'])->name('estados_tecnicos.store');
+Route::get('/estados_tecnicos/{estado_tecnico}/edit', [EstadoTecnicoController::class, 'edit'])->middleware(['auth', 'verified', 'rol.admin'])->name('estados_tecnicos.edit');
+Route::put('/estados_tecnicos/{estado_tecnico}', [EstadoTecnicoController::class, 'update'])->middleware(['auth', 'verified', 'rol.admin'])->name('estados_tecnicos.update');
+Route::delete('/estados_tecnicos/{estado_tecnico}', [EstadoTecnicoController::class, 'destroy'])->middleware(['auth', 'verified', 'rol.admin'])->name('estados_tecnicos.destroy');
